Cosmetic changes to keep Eclipse happy
git-svn-id: https://svn.apache.org/repos/asf/xmlgraphics/fop/branches/FOP_0-20-0_Alt-Design@197442 13f79535-47bb-0310-9956-ffa450edef68

+++ src/java/org/apache/fop/fonts/MultiByteFont.java
@@ -200,7 +200,7 @@ public class MultiByteFont extends CIDFont {
      * @see org.apache.fop.fonts.Font#mapChar(char)
      */
     public char mapChar(char c) {
-        int idx = (int)c;
+        int idx = c;
         int retIdx = 0;
 
         for (int i = 0; (i < bfentries.length) && retIdx == 0; i++) {
